Unified interface, implemented plugin system and added unit tests
By implementing the plugin system, it gives the advantage of ensuring that the user is always using the same keyring backend, prevent unwanted keyring changes during runtime.

Keyring interface has been unified and removed code duplication, specifically the _ensure* methods, now each backend can focus on working with the data instead of checking if it's valid or not since that is done a higher level.

Lastly unit tests were implemented for the base class and the texfile json backend.

VPNLINUX-354