Properly handle unsupported auth_version [VPNLINUX-377]
When the auth_version returned by the API is lower the 3 then this a critical crytpo error as the hash can not be properly calculcated. For this to be properly recognized as a crypto error, it has to be raised as such, which up to this point was only raised as ValueError from the original method.