[VPNLINUX-818] Fix invalid modulus error when logging in
It seems that some users experience this error when the ~/.gnupg
directory is not readable/writable. In this case, python-gnupg fails
silently when trying to import the modulus key in the user's keyring
and then the validation of the modulus signature fails during the
login flow.

To avoid this issue, a separate gnupg keyring stored in the
Proton's cache directory is used instead .