[VPNLINUX-851] Sanitize DNS response
Validates the format of the domain obtained from a DNS TXT record
before using it as the alternative routing domain.

The goal is to avoid potential SSO redirects that could
happen if the domain retrieved from the TXT record contains a
full URL, for example the domain plus a path appended by
the attacker.