[VPNLINUX-758] Make alternative routing errors less verbose
We often get long stacktraces when we are not able to retrieve
alternative routes. The error is logged by asyncio because some
async tasks are ignored, meaning that they are not cancelled and
its result is not retrieved either.

This commit makes sure that the alternative routing async tasks are
not ignored to avoid these verbose asyncio errors.